optimisation + new features
Добавлена возможность автоматического переворота экрана после хода (hotkey A).
Внесены небольшие оптимизации.

+++ ChessBoard.cpp
@@ -5,15 +5,15 @@ ChessBoard::ChessBoard(string FENBoard, bool colorToMove) {
     initBoard(FENBoard);
     moves += colorToMove;
 }
-map<char, int> ChessBoard::FENToEnum = {{'K', WHITE_KING}, {'k', BLACK_KING}, {'Q', WHITE_QUEEN}, {'q', BLACK_QUEEN},
+map<char, int8_t> ChessBoard::FENToEnum = {{'K', WHITE_KING}, {'k', BLACK_KING}, {'Q', WHITE_QUEEN}, {'q', BLACK_QUEEN},
                                  {'R', WHITE_ROOK}, {'r', BLACK_ROOK},{'B', WHITE_BISHOP}, {'b', BLACK_BISHOP},
                                  {'N', WHITE_KNIGHT}, {'n', BLACK_KNIGHT}, {'P', WHITE_PAWN}, {'p', BLACK_PAWN}};
 void ChessBoard::initBoard(string FENBoard) {
-    for(int i = 0, k = 0; i<FENBoard.length(); ++i) {
+    for(int8_t i = 0, k = 0; i<FENBoard.length(); ++i) {
         if (isdigit(FENBoard[i])) {
             k += FENBoard[i] - '0';
         }
-        int x = (k & 7), y = k >> 3;
+        int8_t x = (k & 7), y = k >> 3;
         auto type = FENToEnum.find(FENBoard[i]);
         if (type == FENToEnum.end()) {
             continue;
@@ -28,7 +28,7 @@ void ChessBoard::initBoard(string FENBoard) {
         ++k;
     }
 }
-bool ChessBoard::move(int x, int y, int toX, int toY) {
+bool ChessBoard::move(int8_t x, int8_t y, int8_t toX, int8_t toY) {
     inMove = 1;
     if (canMove(x, y, toX, toY)) {
         board[toX][toY] = board[x][y];
@@ -52,7 +52,7 @@ bool ChessBoard::move(int x, int y, int toX, int toY) {
     inMove = 0;
     return 0;
 }
-bool ChessBoard::canMove(int x, int y, int toX, int toY) {
+bool ChessBoard::canMove(int8_t x, int8_t y, int8_t toX, int8_t toY) {
     //auto t = cached.find({ x, y }); ломает рокировку, canMoveTo не вызывается при inMove
     //if (t != cached.end())
         //return t->second[toX][toY];
@@ -61,7 +61,7 @@ bool ChessBoard::canMove(int x, int y, int toX, int toY) {
     }
     if (board[x][y].canMoveTo(toX, toY)) {
         Figure t = board[toX][toY], t1 = board[x][y];
-        int KX, KY;
+        int8_t KX, KY;
         board[toX][toY] = board[x][y];
         board[x][y].reset();
         board[toX][toY].updatePos(toX, toY);
@@ -90,10 +90,10 @@ bool ChessBoard::move(string from, string to) {
     Square t = ANToXY(from), t1 = ANToXY(to);
     return move(t.x, t.y, t1.x, t1.y);
 }
-bool ChessBoard::isUnderAttack(int x, int y, int ignoreColor) {
-    int color = (!board[x][y]) ? (ignoreColor == -1 ? 42 : ignoreColor) : board[x][y].color;
-    for(int i = 0; i<8; ++i) {
-        for(int j = 0; j<8; ++j) {
+bool ChessBoard::isUnderAttack(int8_t x, int8_t y, int8_t ignoreColor) {
+    int8_t color = (!board[x][y]) ? (ignoreColor == -1 ? 42 : ignoreColor) : board[x][y].color;
+    for(int8_t i = 0; i<8; ++i) {
+        for(int8_t j = 0; j<8; ++j) {
             if (board[i][j] && board[i][j].color != color && board[i][j].canMoveTo(x, y)) {
                 return 1;
             }
@@ -104,15 +104,15 @@ bool ChessBoard::isUnderAttack(int x, int y, int ignoreColor) {
 bool ChessBoard::isKInCheck(bool color) {
     return isUnderAttack(color ? BKX : WKX, color ? BKY : WKY, color);
 }
-void ChessBoard::generatePossibleMoves(int x, int y) {
+void ChessBoard::generatePossibleMoves(int8_t x, int8_t y) {
     if (cached.find({ x, y }) != cached.end()) {
         return;
     }
     vector<vector<bool>> result(8);
     for (auto& i : result)
         i.resize(8);
-    for (int i = 0; i < 8; ++i)
-        for (int j = 0; j < 8; ++j) {
+    for (int8_t i = 0; i < 8; ++i)
+        for (int8_t j = 0; j < 8; ++j) {
             result[i][j] = canMove(x, y, i, j);
             if (result[i][j]) {
                 ++cachedMoves;
